Refactor: move HttpClientFactory internals to .Internal
Moved ActiveHandlerEntry and ExpiredHandlerEntry to separate files under Assimalign.Cohesion.Http.Internal. Relocated all internal implementation details, including HttpClientFactory, to the .Internal namespace. Updated usings, namespaces, and test references accordingly. No functional changes; improves code organization and encapsulation.
