+++ libraries/Http/Assimalign.Cohesion.Http.ClientFactory/src/HttpClientFactoryOptions.cs
@@ -1,3 +1,4 @@
+using Assimalign.Cohesion.Http.Internal;
 using System;
 using System.Collections.Generic;
